CxSAST tests
