+++ tests/CxSAST/CxRestAPI/test_osa_api.py
@@ -45,7 +45,7 @@ def test_create_an_osa_scan_request():
     project_id = get_project_id()
     osa_api = OsaAPI()
     parent_folder = dirname(__file__)
-    path = normpath(join(parent_folder, "../JavaVulnerableLab-master.zip"))
+    path = normpath(join(parent_folder, "../../JavaVulnerableLab-master.zip"))
     scan_id = osa_api.create_an_osa_scan_request(project_id, zipped_source_path=str(path))
     assert scan_id is not None
